+++ app/code/local/PayU/Account/Block/Redirect.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
-*	ver. 1.8.1
+*	ver. 1.8.2
 *	PayU Payment Redirect Block
 *   Payment
 *